+++ src/lib/types/twilio/webhook-payload-reiceved.type.ts
@@ -1,0 +1,14 @@
+export interface TwilioWebhookPayload {
+  Body: string;
+  From: string;
+  To: string;
+  WaId?: string;
+  ProfileName?: string;
+  MessageSid: string;
+  AccountSid: string;
+  NumMedia?: string;
+  MessageType?: string;
+  ChannelMetadata?: string;
+
+  [key: string]: string | undefined;
+}